project/requirements/base.txt
crispy-bootstrap5==2024.2
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-minio-storage==0.5.7
django-split-settings==1.3.2
django==5.0.8
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0